Some refactoring of features summarizer
